+++ api.go
@@ -1,0 +1,32 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+)
+
+type ApiConfig struct {
+	fileserverHits int
+}
+
+func (cfg *ApiConfig) incrementHits(next http.Handler) http.Handler {
+	incr := func(w http.ResponseWriter, r *http.Request) {
+		cfg.fileserverHits++
+
+		next.ServeHTTP(w, r)
+	}
+
+	return http.HandlerFunc(incr)
+}
+
+func (cfg *ApiConfig) getHits(w http.ResponseWriter, r *http.Request) {
+	w.Header().Add("Content-Type", "text/plain; charset=utf-8")
+	w.WriteHeader(200)
+	w.Write([]byte("Hits: " + fmt.Sprint(cfg.fileserverHits)))
+}
+
+func (cfg *ApiConfig) resetHits(w http.ResponseWriter, _ *http.Request) {
+	cfg.fileserverHits = 0
+	w.Header().Add("Content-Type", "text/plain; charset=utf-8")
+	w.WriteHeader(200)
+}
